requirements.txt
playwright
requests
pytest
pytest-playwright
allure-pytest
